+++ source/MaterialXRuntime/Private/Commands/PvtRemovePrimCmd.h
@@ -6,7 +6,7 @@
 #ifndef MATERIALX_PVTREMOVEPRIMCMD_H
 #define MATERIALX_PVTREMOVEPRIMCMD_H
 
-#include <MaterialXRuntime/Private/PvtCommandEngine.h>
+#include <MaterialXRuntime/Private/PvtCommand.h>
 
 #include <MaterialXRuntime/RtStage.h>
 #include <MaterialXRuntime/RtPrim.h>
