+++ source/MaterialXRuntime/Private/PvtCommand.h
